requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
asttokens 
astunparse==1.6.3
attrs==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama 
comm
contourpy==1.3.1
cycler==0.12.1
debugpy 
decorator 
exceptiongroup 
executing 
fastapi==0.116.1
fastjsonschema==2.21.1
flatbuffers==25.2.10
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
idna==3.10
huggingface_hub==0.34.4
importlib_metadata 
ipykernel 
ipython 
jedi 
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client 
jupyter_core 
keras==3.11.1
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline 
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
nbformat==5.10.4
nest_asyncio 
numpy==2.2.2
opt_einsum==3.4.0
optree==0.17.0
packaging 
pandas==2.2.3
parso 
pickleshare 
pillow==11.1.0
platformdirs 
prompt_toolkit 
protobuf==6.31.1
psutil 
pure_eval 
pydantic==2.11.7
pydantic_core==2.33.2
Pygments 
pyparsing==3.2.1
python-dateutil 
pytz==2024.2
pyzmq 
referencing==0.36.2
requests==2.32.4
rich==14.1.0
rpds-py==0.22.3
setuptools==80.9.0
six 
sniffio==1.3.1
split-folders==0.5.1
stack_data 
starlette==0.47.2
tenacity==9.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0rc0
termcolor==3.1.0
threadpoolctl==3.5.0
tornado 
traitlets 
typing-inspection==0.4.1
typing_extensions 
tzdata==2024.2
urllib3==2.5.0
uvicorn ==0.30.1
wcwidth 
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
zipp 
